move related test and remove function from color submodul
